+++ web/src/components/ResetPassword/ResetPassword.test.tsx
@@ -1,0 +1,47 @@
+import userEvent from '@testing-library/user-event'
+
+import { render, screen, waitFor } from '@redwoodjs/testing/web'
+
+import { ResetPassword } from './ResetPassword'
+
+jest.mock('@redwoodjs/web/toast', () => ({
+  ...jest.requireActual('@redwoodjs/web/toast'),
+  success: jest.fn(),
+  Toaster: () => <div></div>,
+}))
+
+const mockReset = jest.fn().mockResolvedValue({})
+
+jest.mock('@redwoodjs/auth', () => ({
+  ...jest.requireActual('@redwoodjs/auth'),
+  useAuth: () => ({
+    isAuthenticated: false,
+    validateResetToken: jest.fn().mockResolvedValue({}),
+    resetPassword: mockReset,
+    reauthenticate: jest.fn(),
+  }),
+}))
+
+const renderComponent = () =>
+  render(
+    <ResetPassword resetToken={'foo'} title={''} message={'New Password'} />
+  )
+
+describe('Reset Password Page', () => {
+  it('renders without crashing', () => {
+    expect(renderComponent).not.toThrow()
+  })
+
+  // it('sends updated password with token for reset', async () => {
+  //   renderComponent()
+  //   const passwordInput = screen.getByLabelText('New Password').closest('input')
+  //   await waitFor(() => userEvent.type(passwordInput, 'supersecret'))
+
+  //   const submitButton = screen.getByRole('button', { name: 'Submit' })
+  //   await waitFor(() => userEvent.click(submitButton))
+  //   expect(mockReset).toHaveBeenCalledWith({
+  //     resetToken: 'foo',
+  //     password: 'supersecret',
+  //   })
+  // })
+})
