+++ web/src/pages/CreatePasswordPage/CreatePasswordPage.stories.tsx
@@ -1,0 +1,7 @@
+import CreatePasswordPage from './CreatePasswordPage'
+
+export const generated = (args) => {
+  return <CreatePasswordPage {...args} />
+}
+
+export default { title: 'Pages/CreatePasswordPage' }
