+++ web/src/pages/CreatePasswordPage/CreatePasswordPage.test.tsx
@@ -1,0 +1,11 @@
+import { render } from '@redwoodjs/testing/web'
+
+import CreatePasswordPage from './CreatePasswordPage'
+
+describe('CreatePasswordPage', () => {
+  it('renders successfully', () => {
+    expect(() => {
+      render(<CreatePasswordPage resetToken={'foo'} />)
+    }).not.toThrow()
+  })
+})
